third_party/WebKit/LayoutTests/fast/css/invalidation/targeted-class-any-pseudo-expected.txt

Check that targeted class invalidation works with :-webkit-any selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(t1, null).backgroundColor is transparent
PASS getComputedStyle(b2, null).backgroundColor is transparent
PASS getComputedStyle(b3, null).backgroundColor is transparent
PASS getComputedStyle(c3, null).backgroundColor is transparent
PASS getComputedStyle(spans4[i], null).backgroundColor is transparent
PASS getComputedStyle(spans4[i], null).backgroundColor is transparent
PASS getComputedStyle(spans4[i], null).backgroundColor is transparent
PASS getComputedStyle(spans4[i], null).backgroundColor is transparent
PASS getComputedStyle(d5, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(t1, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(b2, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(b3, null).backgroundColor is green
PASS getComputedStyle(c3, null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS getComputedStyle(spans4[0], null).backgroundColor is green
PASS getComputedStyle(spans4[1], null).backgroundColor is transparent
PASS getComputedStyle(spans4[2], null).backgroundColor is transparent
PASS getComputedStyle(spans4[3], null).backgroundColor is green
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(d5, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
